Add storybook example
